[kernel][phys] Leave AddressSpace incomplete in the EFI toolchain
Our EFI toolchain does not like large data alignments and in particular
rejects alignments at 16KiB. The use of hwreg::AlignedTableStorage in
AddressSpace thus results in a compilation error in this environment
when we update code to statically specify 16KiB pages on arm64. Since
AddressSpace itself is not relevant at all to the EFI environment, we
address this issue by keeping AddressSpace an incomplete type and out
of the build graph in that case.

Bug: 42107211
Change-Id: I8cb67294ce45960b52f819f225ff134e061c92c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1398596
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>