[reachability] Improve local calculation
Previously, when routes or neighbors were discovered, the entire system
IPv4 and IPv6 were set to at least Gateway state. This is not
necessarily correct since there may have been a v4 route to a neighbor,
but no v6 routes present. Improve this calculation to check for the
protocol of the routes and the neighbor routes.

Fixed: 447422617
Change-Id: Ic15d50dc8371c61f4733d372937a2faa2c10cb26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1374403
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>