[antlion] Fix DictProxy error in WMM test
The WMM tests did not run because of a DictProxy subscript error when
accepting a DictProxy parameter. This might have been a side-effect of
moving in-tree.

The test only ever used the same two DictProxy fields as parameters.
Change the function taking the parameter to directly use the field
instead.

Change-Id: I04b0b648854bb52d54effb26788a3f362d891f6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1396377
Reviewed-by: Karl Ward <karlward@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>