[starnix][mm] Close off Mapping::mapping_name
This does not need to be pub, it seems.

It is also a good idea to close it off because the `Mapping` type
is sensitive to size changes.

Added docs while I was at it. It is important for the properties
of this type to be present in rustdocs for efficient use.

Bug: 422533641
Change-Id: Iffc1548ca633fddc8b67760871d91a79ee1ac079
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1395491
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Maria Glukhova <mariagl@google.com>