[power][drivers] Add more inspect to wake leases
Add the ability to specify a long lease threshold and record
data in inspect when the threshold is exceeded. Adds counting
for when the lease was requested for "too" long and when the
wake lease object held an SAG wake lease "too" long.

Bug: 438827940

Change-Id: I3235a0114c281f043ed0ab9efa7243eefaf8fa7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1398594
Commit-Queue: Justin Mattson <jmatt@google.com>
Fuchsia-Auto-Submit: Justin Mattson <jmatt@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>