[kernel][riscv] Do not reference unified aspace after it's destroyed
When destroying a shared process, the unified aspace gets destroyed
before the restricted aspace is destroyed. When destroying the restricted
aspace, the unified aspace is referenced when flushing ASIDs, which
leads to UB.

On destruction of the unified aspace, we clear the restricted aspace's
reference to the unified aspace and check that the reference is non-null
before accessing it.

Fixed: 452972447
Tested: The remaining failing tests for bringup.riscv64-kernel-lto-cfi
now pass
Change-Id: I419d8d40dcfdc5ed598fc1237a46cd17a63001aa

Change-Id: I27bc93f5827239dddba4554e66f1f71cbd9f3078
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1400539
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>