[orchestrate] Create an empty repository for package serving
Previously, orchestrate would start a package server from the product
bundle and user-built packages would be published directly to the
product bundle. This mostly worked, but only because the internal
structure of a product bundle was similar to that of a package repo.
However, product bundles are not designed to be published to, and this
may break in the future.

This CL changes the flow to first create an empty package repository,
then publish all of the packages from the product bundle into it, and
then finally publish any user-built packages.

Bug: b/334950748
Change-Id: Icbd8c5647264351a0a54bbc87efc997a31787edd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028621
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Michael Brase <mbrase@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>