[netstack3] Make address matchers more type-safe
As written, there was no type-level guarantee that the extension types
for subnet and address range matchers were valid, which created a need
for handling some unnecessary cases. This change bakes those assumptions
into the types themselves.

The subnet matcher was already using net_types::ip::Subnet for
validation, so this change just makes the use explicit (though with
SubnetEither, to not expose the address type as a generic parameter).

Also create a new FIDL-to-core conversion for AddressMatcherEither and
use that as an intermediary in the filtering address matcher conversion.

Change-Id: I0d8e54dcbbff3e6abe61b8531cfee3886a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1400635
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>