[ddk] Avoid host-test deps on some Fuchsia-only code
The ddk, ddktl, and driver_runtime libraries are reached by test
code built for host platforms, not only for Fuchsia.  But some of
the deps of these such as libsync, can never be used on host.  As
such, they aren't actually linked into host code in reality.  But
the deps paths cause them to get compiled, and this happens to
work on Linux but keeping it working constrains refactors in code
never meant to be compiled for non-Fuchsia.  Conditionalize some
deps paths so this Fuchsia-only code won't be compiled elsewhere.

Change-Id: Ia07e01a1f0f42dcaa49bbf2ce590e56d3a28bc84
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1400653
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>