[ffx] Convert target list to use TargetInfo
Change the implementation of `ffx target list` (specifically
target_formatter) to use the new TargetInfo struct.

Note that this also means that "Disconnected" targets will now be
displayed as "Unknown". While this is a functional regression when using
the daemon, in reality, "Disconnected" targets were  more of a source of
problems than of any benefit. In direct mode, we would never be aware of
disconnected targets anyway.

Bug: 450619066

Test: Updated unit tests

Change-Id: I27d0f505d3929e0a461e285be6bda7826a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1400694
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>