[bt][init] Use config capability for Fast Pair
bt-init dynamically routes the `sys.Pairing` FIDL capability to
bt-fastpair-provider if it is available in the system. It currently
relies on package resolving to determine whether the child component is
included in the build or not.

Define a new configuration capability for the Fast Pair Provider feature
and use it in bt-init to determine whether to start bt-fastpair-provider
or not. We will no longer see any package resolving errors which often
cause confusion when debugging.

Update the core realm to route the `fuchsia.bluetooth.FastPairProvider`
config capability to the system. There is no need for a new product
assembly schema since this component is not currently used by any
products.

Bug: 452019054
Test: fx build, fx ota, verify no pkg-resolver logs
Change-Id: Ic1990720eac646223fd64079da0ccfae4b936585
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1392593
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Gary Bressler <geb@google.com>