[blobfs] Refactor how verification works with transfer buffers
This change makes it so that the way in which pages are transferred
prior to verification is consistent between compressed and uncompressed
blobs. Prior to this change, we would verify pages for uncompressed
blobs in the VMO that the driver has access to, which makes them
vulnerable to manipulation if the driver is compromised. This might not
make that much of a difference if the driver has full access to all
memory (which is sometimes the case), but this change will still reduce
an attacker's options.

Bug: 452745864
Change-Id: I1e10dc123eb5b2b84ab37f02d5481b4f74f0363e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1398793
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
API-Review: Chris Suter <csuter@google.com>