[fshost] Use content sniffing to determine when to provision Fxfs
We can use content sniffing to determine if Fxfs should be provisioned.

Doing so has one advantage - we will no longer be reporting expected
errors when attempting to provision Fxfs. Prior to this change, when the
`provision_fxfs` config is set, fshost attempts to mount fxblob and blob
volumes to detect unprovisioned partitions. This failure is expected.
However, the failure will be reported by fshost and logged as an error.

Now, we will only attempt to provision Fxfs if the disk is reported to
not be Fxfs. And will only attempt mounting fxblob and blob volume after
calling provision_fxfs or if disk format is reported as Fxfs.

Update the test to not expect error log messages.

Fixed: 439942311
Test: fx test fshost_integration_tests_storage_host_fxfs
Change-Id: I8f59b4ab248b02ea3306209c8b214c60ae8199aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1393353
Commit-Queue: Rui Qi Sim <rqsim@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>