[debugger][ProcessInfoIterator] Apply recursive filters
This is very important for enabling TestManager to request backtraces of
everything within a particularly failing test realm when the DebugAgent
is recursively attached to all realms ending with `test:test_root`.

When something goes wrong in one of those realms, we only want to get
backtraces from at most all of the processes within that one particular
realm, and at least a process that is crashing within that realm.

Bug: 303314982
Change-Id: Ie02f83cf8134ec849299c2f3338497ae6a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1401953
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>