[block_server] Add support for decompression
Add support for decompression to the block protocol. This will
give us a speed improvement for Blobfs because it won't have to
use the external decompressor.

Also, if we want to support responding to page requests in the driver we
will need this. This is something we are exploring but we haven't
committed to it yet.

Bug: 452745864
Bug: 450908497
Change-Id: Ic598e26510a40c33f23a954759d5b905606a1290
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1393354
Commit-Queue: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>
API-Review: Chris Suter <csuter@google.com>