[ffx][target] Introduce `log-message` command
This CL adds a `log-message` command, which can be used to write an
arbitrary message to the device's log. This is useful when running
manual tests, to give an indication of the important time periods to
readers of the logs.

Test: fx test //src/developer/ffx/plugins/target/log-message:ffx_target_log_message_test
Test: ffx target log-message -s warn -t tag "message"
Change-Id: Ie70365b037e2021e9177338d46f5789c2350e1f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1402455
Commit-Queue: Marc Khouri <mnck@google.com>
Fuchsia-Auto-Submit: Marc Khouri <mnck@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>