[tools] Specify absolute paths for rust-analyzer MCP
Use `$FUCHSIA_DIR` to make the path for the mcp-language-server and
rust-analyzer absolute in the gemini MCP config. This fixes issues where
loading `gemini-cli` in a directory that is not `$FUCHSIA_DIR` would
lead to a MCP initialization error.

A slightly more correct approach would use `${extensionPath}` instead
but that is currently broken for extensions that are installed with
`extension link`.

Fixed: b/454728841
Test: `gemini What is 2 + 4`
Change-Id: I228ee32edd4da7d38739bc99fe62293ac3789385
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1402376
Fuchsia-Auto-Submit: Eric Rahm <erahm@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Eric Rahm <erahm@google.com>