[fxfs] Efficiently clean dirty pages when under memory pressure
Fxfs blocks ZX_PAGER_VMO_DIRTY requests if the system is under critical
memory pressure and more than 1% of system memory is consumed by dirty
pages. The dirty request has to wait for existing dirty pages to be
cleaned. Cleaning dirty pages was done with the flush_all_files method.
Flushing a file has to check if the metadata is dirty which involves the
zx_pager_query_vmo_stats syscall. For cleaning dirty pages, we only care
if a file has dirty pages and not if the metadata is dirty. The
minimize_memory method only flushes files with dirty pages and we know
the number of dirty pages in a file within fxfs so this is much more
efficient.

Fixed: b/421199208
Change-Id: I9140f2c885e773b67796be2b17bf661378275855
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1395195
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>