Revert "[component_manager] disable flaking test"
This reverts commit 979589d9a31e4761037b3eca43d4fcb3ecfb5708.

There was a timing issue in this test. One thread is started with its
own async executor, and that executor is used for the execution scope
with which components are started. The original thread sets up some
components, starts one of them, and signals the second thread to pause
its async executor, effectively pausing any tasks running in the
execution scope of the components.

Once the signal is sent, the original thread does some additional work
and then waits for confirmation from the second thread that it is no
longer running the executor. This additional work on the original thread
causes tasks to be spawned on the execution scope, which then race with
the second thread suspending execution of that scope. In rare cases
(which are more likely under high load), this can cause the test to
fail.

The solution to this issue is simple: wait on the original thread for
the second thread to signal that it's suspended execution on the async
executor before doing any additional work.

Fixed: 453638611
Multiply: component_manager_src_model_tests
Change-Id: I9d6e0299ad8bd4af9c3dd1fd1580d04397d6f867
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1400273
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Olivia Nordquist <nolivia@google.com>