[fidl][rust] Remove Encodable and EncodeRef traits
By switching Encode and EncodeOption to take a wire type parameter as
well as the encoder, these traits can be removed entirely. This also
lays the groundwork for allowing a single natural type to encode as
multiple different wire types depending on the situation.

Change-Id: Ib556e5a5fda3ae062bb7e0874c613d6dde8dcd35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1396953
Commit-Queue: David Koloski <dkoloski@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>