[usb_vsock] Properly handle multiple producers of vsock packets
We were previously only registering one waker for the vsock side
of the `UsbPacketFiller`. This changes it so that we keep a
`Vec` of interested wakers and wake them all when a new
usb packet buffer is registered. This should fix hanging
connection issues when there are lots of vsock connections
on one usb connection.

Fixed: 455634403
Change-Id: Id864e06c69ab6f72d1b1310ca72f01f912aa6781
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1404635
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>