[bazel][idk] Limit valid values for the category attribute
Add `values` for the `category` attribute to IDK macros and rules to
automatically check that valid values are passed.

Bazel does not appear to be enforce values for symbolic macros.

NO_IFTTT=This is a Bazel-specific feature that does not need to be reflected in the GN templates.

Change-Id: I5cad5065735b9b56310be07f7e89ed7f430c5260
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1402855
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>