[assembly] de-dup firmware in OTA manifest
We've had to add two copies of firmware to flash A and B slots, but in
the OTA package the images are identical since the only different - the
flash partition name - isn't used for OTAs.

Add logic to de-dup these images in the OTA manifest, otherwise
downstream consumers of the manifest can have problems.

Bug: b/434935097, b/450918051
Change-Id: Id03cc04f9b491b666d57da89fe50f71d75400cc3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1388933
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Commit-Queue: David Pursell <dpursell@google.com>