[starnix][ptrace] Implement PTRACE_POKEDATA/TEXT
These Ptrace ops allows a tracer to modify memory of a tracee, and works
for private memory that neither tracer nor tracee has write capabilities
to.  This is useful for modifying text (code) regions of a program to
set a software breakpoint.

The semantics of this operation are explored in PtracePokeMemory.*
syscall tests added in previous CLs. See PtraceTest.PokeToSetBreakpoint
for example of how a Linux debugger like LLDB would set a breakpoint in
a running program.

To write to VMOs that Starnix does not hold writable handles to, use
zx_vmo_create_child to create a copy-on-write child, and remint the VMO
as executable if the original VMO was executable.

On x86_64, the PC register is advanced after a breakpoint instruction,
but this is not the case on other architectures. Updated the test to
reflect this. Also added padding to previously added test so that ptrace
does not overwrite the next instruction.

Bug: 445790710
Change-Id: I4f224c6319ad5aaa3bf0cc73705b831b245aa08f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1381185
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: David Song <wintermelons@google.com>