[starnix] Remove message from StubEmptyFile
StubEmptyFile only ever logs its path as a message. To prevent needing
to pass a path to StubEmptyFile on creation (which might not even be
right, considering that a StubEmptyFile can be mounted in different
namespaces), add an `open` method to FileOps and call it when the
FileObject is created which corresponds to the union of a Namespace
and a FileOps.

In StubEmptyFile, implement the open method to log the path that the
file was opened with.

This log was previously emitted at create_file_ops time, which
corresponds to when the file was opened anyway, so there is no
behavior change by logging on open.

Test: existing tests
Change-Id: Ia18e5ea38991982f9fc6fa6a5e3eec832d1e931d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1400553
Reviewed-by: Theodore Dubois <tbodt@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: John Wittrock <wittrock@google.com>
Commit-Queue: John Wittrock <wittrock@google.com>