[cml-validate] validate facets
original algorithm assumes all facets are merged together
and makes a validation assessment of the full set.

now we're going document by document so it's possible that
individually defined facets could be invalid but the superset
would be valid. this can happen if both
EnableAllowNonHermeticPackagesFeature AND AllowNonHermeticPackages
are set to true but a facet in a document doesn't contain
TEST_DEPRECATED_ALLOWED_PACKAGES_FACET_KEY but an included document
does have the key.

We can avoid this false negative by deferring the check until we have
seen all of the facets that transitively form the superset of facets.
If there still isn't a TEST_DEPRECATED_ALLOWED_PACKAGES_FACET_KEY then
we can return the validation error. It's not a spanned error because
the error is with the Features instead of with any Document.

Change-Id: I17a42510e95247841c783d40fd7a258fbdd7fcb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1404515
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Olivia Nordquist <nolivia@google.com>