[rust-3p] Update README for ask2patch
This adds an additional step in the procedure to revert the temporary
changes to Cargo.lock and BUILD.gn that pile up after the intermediate
step of running `fx update-rustc-third-party`. This avoids ending up
with the same version of the crate in ask2patch/ and vendor/ and the
subsequent need to clean up the latter manually.

Change-Id: I9c98a3bc5d2593bf2e1440659cc216e8fd1b0bf6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1402275
Reviewed-by: David Koloski <dkoloski@google.com>
Commit-Queue: Thomas Zander <thomasalpinus@google.com>