[usb][dwc3] Start soft transition to managing vbus valid signal
dwc3 must manage the vbus valid signal on platforms that support it.
This change adds a check for the MMIO size to determine whether or not
the vbus valid register is accessible, and if it is, sets or clears the
signal as needed. This check will be removed after the soft transition
has been completed.

Bug: 450597235
Test: Booted Sorrel, USB worked on boot and after disconnect/reconnect
Change-Id: I37dbc98cc1faa98619ec519a55c2bb3c079ca7d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405493
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Fuchsia-Auto-Submit: Braden Kell <bradenkell@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>