[sestarnix] Integrate LSM checks with default_check_access()
LSM's fs_node_permission() hook is intended to be called both when
a file is being opened and for any other calls that perform
discretionary access checks.

The hook is now integrated into the default_check_access()
implementation so that LSM is applied to these checks.

The hook remains a no-op unless one of the ACCESS or FOR_OPEN flags
is set in the requested PermissionFlags - the other cases will be
integrated as a todo-check, before being migrated to be fully
enforced.

Bug: 364568874, 455783684
Change-Id: Id8c275fcb1240934efbe499ce39758a3c67070ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1334204
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Wez <wez@google.com>