[kernel][vm] Add a subregion lock to VmAspace
Introduces a new lock to VmAspace that protects the subregion tree.
Both the general VmAspace lock, and the new region lock, must be held
to modify the tree, but either lock can be used to walk the tree.

Having this separate lock will allow for operations that are only
iterating the tree, namely AllocSpotLocked, to operate without holding
the general aspace lock. This allows non allocation operations, such as
page faults, to happen in parallel.

As of this commit the regular lock_, in addition to the new region
lock, is held over spot allocation, and as such this commit has no
behavioral changes.

Change-Id: I8a3023cb3989ac437bfba2d895fa88bb24f930c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364814
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>