[ffx] Fix #[target(direct)]
The FfxTool attribute #[target(direct))] was a no-op, due to mishandling
of state in the FhoTargetEnvironment. The entire thing needed to be
wrapped in an Arc so that changes so that when retrieved, any changes
(specifically to "want_direct") would be persisted.

Test: Added unit test to validate the correct behavior, and manually
confirmed that the only user, `ffx trace`, correctly uses direct mode.

Fixed: 453670791

[g-review]

Change-Id: I7747cf5ba2ce79dc97ff7f989e02ca5b6a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405279
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>