[cml-validate] even more offers validation!
checks if duplicate protocols or duplicate
dictionaries are offered to all. edits the code
in the original function because i struggled to be
sure it was the same algorithm as the one i implemented.
lo and behold, they are!

also added the yet-unused-CapabilityRequirements
so that the tests are semantically similar. the dead-code
tag should be shortly addressed but the scope was starting
to bloat.

Change-Id: I6f9e47c56837152333d6c7a24d72965c10819c53
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1406173
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Olivia Nordquist <nolivia@google.com>