[utc][starnix][vdso] Use mappable clocks in starnix vdso
Use the mappable clocks Zircon feature to implement UTC clock for
Starnix. This change replaces relying on a "clock modified" signal,
which could cause a mismatch between Fuchsia and Starnix readings of the
UTC clock.

Starnix still uses a simulated UTC clock while Fuchsia's UTC clock is
not started, so that programs under Starnix always observe a moving UTC
clock. But, once Fuchsia's UTC clock is started, Starnix will switch to
reading the real memory mapped UTC clock. These reads are synchronized
with clock updates by the Zircon kernel, so that the UTC clock reads
within Starnix are always consistent with what other processes might
read.

This consistency is important for scenarios where a Fuchsia process must
agree with a process running under Starnix on what the UTC reading is.

One real situation where this is important is the handling of the wake
alarms on the UTC timeline, a.k.a. "realtime alarms" under Starnix. When
a realtime alarm fires due to UTC timeline change, Fuchsia's alarm
manager will signal this to Starnix. But, if Starnix does not observe
the deadline as expired because the notification of timeline change was
not yet processed at time of check, it may ignore the notification, and
continue to block. Since Fuchsia already notified of timer expiry, that
another notification may never come. The starnix process then remains
blocked forever, instead of unblocking on an alarm eventually.

Bug: 422533641, 415099415
Change-Id: I850e0f8b45f8f187b59e86eabfbd6c79e6e56d57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1297350
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: James Robinson <jamesr@google.com>