[build][rust] fx rustdoc-link uses relpath instead of pathlib's relative_to
This change fixes an error when running `fx rustdoc-link`.
The error appears as: "rustdoc_link_verify.py is not in the subpath of"
the build directory.

os.path.relpath does a "walk up" path computation, which is what we need
here. Use it instead of pathlib's relative_to.

Fixes: 455853676
Change-Id: I34cc35149fe65df8dcd339904f11d4bd0fd23372
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405573
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Ethan Williams <ethanws@google.com>