[starnix] Propagate test runner failures.
The current practice of logging test runner errors has resulted in false
positives, due to a bug in the test runner framework. Untimely crashes
in the Starnix test runner were causing zero results to be reported at
all, which in turn the framework was interpreting as "no failures."

While that bug has been fixed, I don't think the test runner simply
logging runtime errors is the best strategy here. Crashing the entire
component via expect is extreme, it's also the only reasonable solution
I can think of in the case that a run request falls over. Given the
context of a test runner, I'd argue it's better to fail loudly if
there's an internal error.

I've got reason to believe that this is more than a hypothetical
problem. Last time I tried to land a very similar change, it resulted in
a huge spike in reported test flakes. That change also introduced
threading changes, which may actually be the root cause. However, the
test runner framework wasn't complaining about timeouts or
unresponsiveness. I'd love to be proven wrong, and to land this change
and nothing happens. But there's a chance the spike in flakes was due
to the test runner eating up runtime errors, and if so I'd like to
verify and get to the bottom of it.

Bug: 438571603
Test: fx test starnix_syscalls_cpp_tests
Change-Id: I9e00f02c71ee9e08d7ca69720243c14b9ca7b976
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1406813
Commit-Queue: Aaron Knobloch <aknobloch@google.com>
Reviewed-by: Maria Glukhova <mariagl@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Aaron Knobloch <aknobloch@google.com>