[kernel][vm] MutatePages for VmPageSpliceList
Add a function to the VmPageSpliceList, MutatePages, that allows a pass
through a finalized VmPageSpliceList to modify pages in the splice list
without transferring ownership or considering them 'processed'.

This is useful in cases where the caller might have to deal with
ZX_ERR_SHOULD_WAIT, as the current VmPageSpliceList iterators transfer
ownership to the caller, and pages can not be re-appended in the
Finalized state.

Change-Id: I30bada5cdf1c89bd05a5b4fe432c2a54693a6cad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1398093
Commit-Queue: Sage Barreda <sagebarreda@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>