[roll] Roll third_party/mesa vulkan/queue: Move shared binary semaphores to temps
If a client creates a semaphore, exports it, and then re-imports it back
into the device, this can trick our semaphore reset logic.  When this
happens, we end up with two different vk_sync structs that have the same
underlying payload so if one is used as the signal and one is used as
the wait of the same submit, we'll end up resetting it because we think
they're different, causing us to lose the signal.

We already have the ability to handle this for the threaded case by
moving the semaphore payload into a new vk_sync which we then destroy
after we're done submitting to the driver.  Use this path for shared
semaphores in the immediate case so we can just wait and signal without
worrying about the reset.

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/13805
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37149>
(cherry picked from commit 0f7e0f79ad1754d64cef08da0a511c8447bb8d44)
Original-Revision: 23495cf441d11d6a570cacfc884b026044025cea

Rolled-Repo: https://fuchsia.googlesource.com/third_party/mesa
Rolled-Commits: 4533272f9cd836..23495cf441d11d
Roller-URL: https://cr-buildbucket.appspot.com/build/8699624763589253409
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I0b55f4c532d11bd79950ea250c677b5535829dfa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1407827