[sestarnix] Add a test for type transitions on FUSE
FUSE filesystems are typically genfscon labeled, and do not support
seclabel. In this case, type transitions are applied to compute a
label and perform may_create checks, but the label is subsequently
discarded in favor of the genfscon label. This test implements a very
simple FUSE filesystem to test this behavior. The test is currently
expected to fail on Starnix.

Bug: 414336570

Change-Id: I662837e1a194c90d9a536adf2a10acbdcada7c02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1398814
Reviewed-by: Marina Ciocea <marinaciocea@google.com>
Commit-Queue: Ambre Williams <ambre@google.com>