[bootfs] Replace `bin/startup` with eager
To make this possible, change eager startup failure behavior to just log the
error instead of propagating the error to the parent's start. This is
necessary because some `core` components in certain builds have a child
definition but their package is not part of the build. This should
probably be fixed, but the new behavior is arguably more conceptually
coherent -- the lifecycles of parent and child instances are normally
decoupled and if the child is started by some other means it doesn't
affect the parent.

The benefit is we can reclaim a bit of memory from bootfs, as well as
simplify the topology slightly.

Bug: 438195589
Change-Id: Id1ebf5af8655d898bba2b34e27243dc88c370943
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1396566
Reviewed-by: Ali Zhang <alizhang@google.com>
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>