[trace] Support ingesting a 'split' JSON trace
The trace_processing library can now import a trace split across two files:
- A valid Chromium JSON-formatted trace, without `systemTraceEvents`
- The content that would be in `systemTraceEvents` in jsonlines format

This reduces the RAM needed to import a trace formatted in JSON, as one
of the main drivers of memory usage during import is the need to hold
the entire JSON version of the trace in memory. The new format allows
process, thread and scheduler records to be streamed through the
importing code.

Bug: 437397276
Test: fx test trace_importing_test

Change-Id: I92e2ded766862353004d5b0fe25e423daa852c62
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1407573
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>