[kernel] Introduce <lib/page/size.h>
This change introduces a simple 'library' meant to consolidate reference
to and dependency on system page size *within* the kernel (with
visibility constraints as well). This preserves the current expectation
that page size is available as a constant, but is intended to break its
relationship to the zircon public headers. Follow-on changes will
replace usage of (ZX_)?PAGE_SIZE (and friends) in the kernel with the
definitions in this header.

Bug: 42146863
Change-Id: I82860fb23360a4c91aaaaeef55127ba6b4c0f8de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405318
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>