[kernel][phys] Prevent ElfImages from being destroyed
In this test, we add a bunch of local ElfImages to the global module
list in gSymbolize. These ElfImages are destroyed at the end of their
respective scopes, but the module list still has pointers to them which
are referenced in Symbolize::module_for_vaddr_range which is used
whenever we would do a cfi check via indirect calls via FILE::Write.

Since we don't really have a way of "unloading" modules and I'm not sure
if we want to expose a method for removing them from the modules list in
Symbolize, this instead makes the ElfImages persistent by wrapping them
with NoDestructor and making them static. Likewise we need to make the
storage for the modules persistent so we make that a global array.

Fixed: 454696559
Change-Id: I619e54127fd85228aba943bbfe339807590fe5bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1402621
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>