[kernel] Replace usage of ZX_PAGE_SIZE
Page size in userspace should be a runtime determination and the export
of page size in <zircon/limits.h> is deprecated. We replace that
deprecated usage with the kernel-internal libpage definition of page
size... with the exception of the last use of ZX_PAGE_SIZE as a default
boot option, to be handled in a follow-on change.

Bug: 42146863
Change-Id: I74fa1abefbf53cc3f29df73029f6662d42752391
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405319
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>