[netlink] Fix IflCacheInfoMissingBody for host
A catastrophic rebase ended up reintroducing the NET_CAP_ADMIN check for
NetlinkRouteTest. Removing that shows that IflaCacheInfoMissingBody was
not actually passing on host.

Massage the test a bit so we still have the coverage with a test that
passes both on starnix and not.

Fixed: 456245996
Change-Id: I29899e4206350b8f78d7324bf6f838ca187c63a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1408575
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Aaron Knobloch <aknobloch@google.com>