[inspect] Make the error message more clear when property match fails
This updates the output to inform the user when they've attempted
to match a property, but the name is actually a node.

We can't/shouldn't use AnyProperty to match empty nodes, because
that breaks the distinction between nodes and properties. But,
we can still suggest the node-equivalent of AnyProperty since
the presence of the node is detectable.

This also updates the test that brought this issue to my attention.

Change-Id: Ie25134cce2cec24b3ad58ff53211d909ece00e0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1408719
Reviewed-by: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Clayton Mccray <claytonmccray@google.com>
Commit-Queue: Clayton Mccray <claytonmccray@google.com>