[ffx] Change `ffx target list` to use is_default
Instead of doing a name comparison, use the TargetInfo::is_default
field when indicating that a target is the default.

Note that _multiple_ devices can be listed as "default". This is
not a regression, since before this CL, multiple devices could be
configured with the same node name, and they would then all be
shown as the "default". In practice, this simply means that
specifying the default device may not actually disambiguate the target,
and ffx target list will reveal that by putting the "*" next to
multiple devices.

(This was surprisingly weird code, because of all the macros in
target_formatter. At some point we'll need to clean all that up.)

Bug: 448164233

Test: updated existing unit tests, and a couple to explicitly test
the new mechanism.

[g-review]

Change-Id: If5c93bf66906ea95b747d1417ed9da0f6a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409458
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>