[tee] Use base resolver for tee_manager
The product must always include this package as part of the base package
set. This is necessary because the full-resolver is unable to resolve
the package when the tee is required to unlock persistent storage,
creating a cycle.

Change-Id: I178abfd7f06fa5f914e8e524883a74996a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405675
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>