[vfs][rust] Remove uses of FutureExt::boxed
Rust doesn't seem to inline trait methods that often. Using
FutureExt::boxed causes the future to be constructed on the stack and
then copied to the heap allocation in the FutureExt::boxed method. When
Box::pin is used directly, rust will make the allocation first and then
construct the future directly on the heap. This avoids the memcpy of
the future.

This optimization is probably only saving nanoseconds and is unlikely to
be detectable in most storage benchmarks.

Bug: b/456517172
Change-Id: I254164df6806fef8cb041fd8db3c48c23e82b19d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409734
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Fuchsia-Auto-Submit: Chris Drouillard <cdrllrd@google.com>